% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBS.findR.R
\name{analyse.pwm}
\alias{analyse.pwm}
\title{analyse.pwm}
\usage{
analyse.pwm(data, pwms)
}
\arguments{
\item{data}{is a single Grangesobject from \code{read.input.file}}

\item{pwms}{is a simplelist which include several (at least 2)
motifs which are position weight matrices, position probabilty matrices or
position frequency matrices.}
}
\value{
GrangesObject with added columns:
\item{Ref.score}{list of PWM compared score for each position
to Ref.sequence}
\item{Alt.score}{list of PWM compared score for each position
to Alt.sequence}
Note that the amount of GrangesObject are also *amount of pwms,
as a score list is provided for each individual pwm
}
\description{
Analyse position weight matrix to Grangesobject from \code{read.input.file}.
This is only to do analyses in loop for each motif in motiflist.
}
