% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.data.R
\name{data.update}
\alias{data.update}
\title{data.update}
\usage{
data.update(data, threshold = 1, pseudocount = 0.01)
}
\arguments{
\item{data}{is a single Granges object from \code{TFBS.findR}}

\item{threshold}{which Delta Kumasaka score should meet.
This will make sure that less relavant changes in motif
scores will not be reported}

\item{pseudocount}{in order to make sure that no -inf values
occur when Kuma score is negative for either ref or alt}
}
\value{
dataframe with the following columns:
\item{seqnames}{Chromsome}
\item{start}{start position of sequence in genome}
\item{end}{end position of sequence in genome}
\item{width}{size of sequence/motif}
\item{strand}{which strand sequence is on}
\item{SNP}{variant rs}
\item{Allel}{Phased allel information}
\item{REF}{Reference nucleotide}
\item{ALT}{Alternative nucleotide}
\item{Snp.loc}{Location of SNP on sequence}
\item{Sequence}{Reference sequence matched with motif}
\item{MotifDB}{Motif found in sequence}
\item{Ref.score}{PWM compared score for Reference}
\item{Alt.score}{PWM compared score for Alternative}
\item{Delta.score}{Alt.score - Ref.score}
\item{Kuma.ref.score}{posterior probability of transcription
factor binding for Reference}
\item{Kuma.alt.score}{posterior probability of transcription
factor binding for Alternative}
\item{Kuma.delta.score}{Kuma.alt.score - Kuma.ref.score}
}
\description{
Updates the data in loop to make snp-motif combination readable
As well as sorting position at which snp changes motif.
Threshold can be used to filter out snp-motif changes
that have low scores, making them less relevant.
Lastly, the data will be transformed from a GRanges object
to a dataframe.
}
