% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBS.findR.R
\name{kumasaka.score}
\alias{kumasaka.score}
\title{kumasaka.score}
\usage{
kumasaka.score(data, background = c(A = 0.25, C = 0.25, G = 0.25, T =
  0.25), prior = 0.1)
}
\arguments{
\item{data}{is a Grangesobject from \code{analyse.pwm}}

\item{background}{is the postion frequency matrix with each
background score of a nucleotide}

\item{prior}{is set the False discovery rate (FDR) which is
set to 10%. 0.1 is the default setting as we normally pass
a dataset from 1 sample.}
}
\value{
Grangesobject with added columns:
\item{Kuma.ref.score}{list of posterior probability of transcription
factor binding score for each position and motif to Ref.sequence}
\item{Kuma.alt.score}{list of posterior probability of transcription
factor binding score for each position and motif to Alt.sequence}
}
\description{
kumaska et al., 2019, used a different method of scoring
in which posterior probability of transcription factor
binding is calculated
MORE INFO
}
\examples{
#data described from analyse.pwm
data <- analyse.pwm(data, motiflist=JASPARCORE, motif.type="PPM")
data <- kumasaka.score(data, background=c(A=0.25, C=0.25, G=0.25, T=0.25), prior=0.1)

}
