% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.value.calculation.R
\name{p.value.calculation}
\alias{p.value.calculation}
\title{p.value.calculation}
\usage{
p.value.calculation(data, motiflist, background = c(A = 0.25, C = 0.25, G
  = 0.25, T = 0.25), motif.type = "PFM", pseudocount = 0.01)
}
\arguments{
\item{data}{is a Grangesobject from \code{update.data.R}}

\item{motiflist}{is a simplelist of motifs}

\item{background}{is the postion frequency matrix with each
background score of a nucleotide}

\item{motif.type}{is the type of motifs passed. These can be
"PWM", "PFM", "PCM" or "PPM".}

\item{pseudocount}{value to change the outcome model, which removes
possibility for -inf values.}
}
\value{
Grangesobject with added columns:
\item{Pvalue.ref.score}{Pvalue of transcription factor binding score
for each position and motif to Ref.sequence}
\item{Pvalue.alt.score}{Pvalue of transcription factor binding score
for each position and motif to Alt.sequence}
}
\description{
Touzet et al., 2007, created a method for calculating
pvalues for PWMs
}
