% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBS.findR.R
\name{pwm.compare}
\alias{pwm.compare}
\title{pwm.compare}
\usage{
pwm.compare(pwm, data)
}
\arguments{
\item{pwm}{are calculated pwm scores from \code{to.PWM}}

\item{data}{are GRanges objects from \code{read.input.file}}
}
\value{
GrangesObject with added columns:
\item{Ref.score}{list of PWM compared score for each position
to Ref.sequence}
\item{Alt.score}{list of PWM compared score for each position
to Alt.sequence}
Note that the amount of GrangesObject are also *amount of pwms,
as a score list is provided for each individual pwm
}
\description{
Compare position weight matrix with both the reference
and alternative sequence in data, in order to calculate
score for each position
}
\details{
Here we analyse our reference and alternative sequence
based on a 'rolling TF' model. Calculated pwm from
\code{to.PWM} is compared to Grangesobject from
\code{read.input.file}. PWM are compared to every possible
position in the sequence, even outside SNP.
}
