% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.input.file.R
\name{read.input.file}
\alias{read.input.file}
\title{read.input.file}
\usage{
read.input.file(input.file, ref.genome, sample.name = "Unknown",
  ATAC.only = FALSE)
}
\arguments{
\item{input.file}{is the path to a file, which can be both a vcf and fasta file.
Totally fine if these are zipped.}

\item{ref.genome}{is the reference genome used to get IRanges and sequence information.}

\item{sample.name}{which is self explanitory, usefull if you want to codense several
outcomes of different samples in a later stage}

\item{ATAC.only}{can be used to pass a bed file which only has the alterations in an
ATAC peak. Based on this bed file, the alterations in input.file will be filtered.}
}
\value{
a GrangesObject with following columns:
\item{Sample}{which is the sample.name passed as param}
\item{SNP}{name of the SNP}
\item{Allel}{Phased allel information if phased vcf is passed}
\item{REF}{Reference nucleotide}
\item{ALT}{Alternative nucleotide}
\item{REF.sequence}{20 nucleotide before and after REF from ref.genome}
\item{ALT.sequence}{20 nucleotide before and after ALT from ref.genome}
}
\description{
Read and strip input file and make ready for motif analysis
}
\examples{
#Load input.files
input <- "/../HG01241.vcf.gz"
ATAC.snps <- '/../ATACpeaksMerged.bed'
#Load reference genome
library(BSgenome.Hsapiens.UCSC.hg19)
ref.genome <- BSgenome.Hsapiens.UCSC.hg19
#Get sample.name from input file
sample.name <- strsplit(input, split="[./_]+")[[1]][2]
data <- read.input.file(input, ref.genome, sample.name = sample.name, ATAC.only = ATAC.snps)

}
