% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.motif.database.R
\name{read.motif.database}
\alias{read.motif.database}
\title{read.motif.database.R}
\usage{
read.motif.database(file)
}
\arguments{
\item{file}{is a path to a .txt file with motif scores}
}
\value{
a SimpleList with motif matrixes
}
\description{
Read and strip text motif database files for motif analysis
}
