% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.plots.R
\name{snp.plot}
\alias{snp.plot}
\title{snp.plot}
\usage{
snp.plot(snp, method = "both", motif = "MEF2A", strand = c("-", "+"))
}
\arguments{
\item{snp}{is a specific GRanges item from TFBS.findR which include only
one snp.}

\item{method}{is the method used for TFBs.findR. if both or Kuma is provided
kumasaka scores will also be plotted.}

\item{motif}{is the motif which should be investigated from the Granges object.
Note that the motif needs to be present in the GRanges object.}

\item{strand}{is the strand which should be investigated from the Granges
Object. As default, both will be accepted, however if both plus and minus
strand are present, than ploting doens't work.}
}
\value{
two or four plots with PWM scores, delta PWM scores, Kumasaka
scores and delta Kumasaka scores
}
\description{
Plots for a specific snp, strand and motif the PWM scores and Kumasaka
scores if provided for Ref vs Alt
}
