% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBS.findR.R
\name{TFBS.findR}
\alias{TFBS.findR}
\title{TFBS.findR}
\usage{
TFBS.findR(data, motiflist, motif.type = "PFM", method = "both",
  background = c(A = 0.25, C = 0.25, G = 0.25, T = 0.25),
  pseudocount = 0.01, prior = 0.1, BPPARAM = bpparam())
}
\arguments{
\item{data}{is a GRangesobject from \code{read.input.file}}

\item{motiflist}{is a simplelist of motifs}

\item{motif.type}{is the type of motifs passed. These can be
"PWM", "PFM", "PCM" or "PPM".}

\item{method}{can be used to determine if kumasaka score will be
calculated.}

\item{background}{is the background probability for each nucleotide.}

\item{pseudocount}{value to change the outcome model, which removes
possibility for -inf values.}

\item{prior}{is set the False discovery rate (FDR) which is
set to 10%. 0.1 is the default setting as we normally pass
a dataset from 1 sample.}

\item{BPPARAM}{are settings by BiocParallel, which can be used
to do multiple core calculations. bpparam() are the current settings
on your system.}
}
\value{
a Grangesobject with following columns:
\item{Sample}{which is the sample.name passed as param}
\item{SNP}{name of the SNP}
\item{Allel}{Phased allel information if phased vcf is passed}
\item{REF}{Reference nucleotide}
\item{ALT}{Alternative nucleotide}
\item{REF.sequence}{20 nucleotide before and after REF from ref.genome}
\item{ALT.sequence}{20 nucleotide before and after ALT from ref.genome}
\item{MotifDB}{which motif database was used}
\item{provider}{provider from the motif}
\item{Motif}{The motif analysed}
\item{Ref.score}{list of PWM compared score for each position
to Ref.sequence}
\item{Alt.score}{list of PWM compared score for each position
to Alt.sequence}
\item{Kuma.ref.score}{list of posterior probability of transcription
factor binding score for each position and motif to Ref.sequence}
\item{Kuma.alt.score}{list of posterior probability of transcription
factor binding score for each position and motif to Alt.sequence}
}
\description{
TFBS.findR analyses Grangesobjects from \code{read.input.file}
if they match motifs in the motiflist.
}
\details{
Motif-sequence analyses is conducted using several steps:
\code{analyse.pwm}. First changes motifs provided in motiflist
to position weight matrices. Secondly, these PWMs are compared
to individual Grangesobjects from \code{read.input.file} and
PWM scores are given for each position in the sequence of these
Grangesobjects
\code{kumasaka.score} analyses values provided by analyse.pwm
and calculates the posterior probability of transcription factor
binding. These values are than similarly passed as a list to the
Granges object
}
