% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBS.findR.R
\name{trim.motifs}
\alias{trim.motifs}
\title{trim.motifs}
\usage{
trim.motifs(motif)
}
\arguments{
\item{motif}{is a single item from motiflist}
}
\value{
trimed specific motifs
}
\description{
Function to trim certain motifs because of inability to calcalate p-value
otherwise
}
\details{
THIS SHOULD BE UNNESCESSARY! however no reaction on git from Ge-Tan
}
