% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.plots.R
\name{variant.plot}
\alias{variant.plot}
\title{variant.plot}
\usage{
variant.plot(snp, MotifDb, method = "both")
}
\arguments{
\item{snp}{is a specific item from TFBS.findR after data.update which include only
one snp.}

\item{MotifDb}{is a simplelist of all the motifs PPM provided within snp.}

\item{method}{is the method used for TFBs.findR. if both or Kuma is provided
kumasaka scores will also be plotted.}
}
\value{
one or two plot for PWM score and/or Kuma score for a specific snp.
}
\description{
Plots for a specific snp, the PWM scores and Kumasaka of all motifs in the variant
}
