% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.data.R
\name{Window.update}
\alias{Window.update}
\title{Window.update}
\usage{
Window.update(data)
}
\arguments{
\item{data}{is a Granges object from \code{GRanges.update}}
}
\value{
dataframe with the following columns:
\item{seqnames}{Chromsome}
\item{start}{start position of sequence in genome}
\item{end}{end position of sequence in genome}
\item{width}{size of sequence/motif}
\item{strand}{which strand sequence is on}
\item{SNP}{variant rs}
\item{Allel}{Phased allel information}
\item{REF}{Reference nucleotide}
\item{ALT}{Alternative nucleotide}
\item{Snp.loc}{Location of SNP on sequence}
\item{Sequence}{Reference sequence matched with motif}
\item{MotifDB}{Motif found in sequence}
\item{Ref.score}{PWM compared score for Reference}
\item{Alt.score}{PWM compared score for Alternative}
\item{Delta.score}{Alt.score - Ref.score}
\item{Kuma.ref.score}{posterior probability of transcription
factor binding for Reference}
\item{Kuma.alt.score}{posterior probability of transcription
factor binding for Alternative}
\item{Kuma.delta.score}{Kuma.alt.score - Kuma.ref.score}
}
\description{
Update data to a dataframe rather than Granges object
}
