% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBS.findR.R
\name{analyse.pwm}
\alias{analyse.pwm}
\title{analyse.pwm}
\usage{
analyse.pwm(data, motiflist, motif.type = "PFM", background = c(A =
  0.25, C = 0.25, G = 0.25, T = 0.25), pseudocount = 0.01)
}
\arguments{
\item{data}{is a single Grangesobject from \code{read.input.file}}

\item{motiflist}{is a simplelist which include several (at least 2)
motifs which are position weight matrices, position probabilty matrices or
position frequency matrices.}

\item{motif.type}{is the type of matrices passed in motiflist. Can be "PFM",
"PCM", "PPM", "PWM".}

\item{background}{is the background probability for each nucleotide.}

\item{pseudocount}{value to change the outcome model, which removes
possibility for -inf values.}
}
\value{
GrangesObject with added columns:
\item{Ref.score}{list of PWM compared score for each position
to Ref.sequence}
\item{Alt.score}{list of PWM compared score for each position
to Alt.sequence}
Note that the amount of GrangesObject are also *amount of pwms,
as a score list is provided for each individual pwm
}
\description{
Analyse position weight matrix to Grangesobject from \code{read.input.file}.
This is only to do analyses in loop for each motif in motiflist.
}
\examples{
#single Grangesobject (single SNP)
data <- read.input.file(path, ref.genome)[1,]
#Here multiple human motifs are provided
library(MotifDb)
JASPARCORE <- query(MotifDb, 'JASPAR_CORE')
JASPARCORE <- query(JASPARCORE, 'hsapiens')
data <- analyse.pwm(data, motiflist=JASPARCORE, motif.type="PPM")

}
