% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.data.R
\name{GRanges.update}
\alias{GRanges.update}
\title{GRanges.update}
\usage{
GRanges.update(data, threshold = 2, pseudocount = 0.01)
}
\arguments{
\item{data}{is a single Granges object from \code{TFBS.findR}}

\item{threshold}{which Delta Kumasaka score should meet.
This will make sure that less relavant changes in motif
scores will not be reported}

\item{pseudocount}{in order to make sure that no -inf values
occur when Kuma score is negative for either ref or alt}
}
\value{
Granges object with columns:
\item{Sample}{which is the sample.name passed as param}
\item{SNP}{name of the SNP}
\item{Allel}{Phased allel information if phased vcf is passed}
\item{REF}{Reference nucleotide}
\item{ALT}{Alternative nucleotide}
\item{Snp.loc}{Location in the motif sequence where variant occurs}
\item{Sequence}{Reference sequence matched with motif}
\item{MotifDB}{which motif database was used}
\item{provider}{which provider name motif is from}
\item{Motif}{The motif analysed}
\item{Ref.score}{PWM compared score for Reference}
\item{Alt.score}{PWM compared score for Alternative}
\item{Delta.score}{Alt.score - Ref.score}
\item{Kuma.ref.score}{posterior probability of transcription
factor binding for Reference}
\item{Kuma.alt.score}{posterior probability of transcription
factor binding for Alternative}
\item{Kuma.delta.score}{Kuma.alt.score - Kuma.ref.score}
}
\description{
Updates the data to make snp-motif combination readable
As well as sorting position at which snp changes motif.
Threshold can be used to filter out snp-motif changes
that have low scores, making them less relevant.
}
